+++ src/main/java/org/codeiteam3/findex/indexinfo/controller/IndexInfoController.java
@@ -5,7 +5,7 @@ import io.swagger.v3.oas.annotations.Parameter;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.codeiteam3.findex.common.CursorPageResponse;
+import org.codeiteam3.findex.pagination.CursorPageResponse;
 import org.codeiteam3.findex.indexinfo.dto.data.CursorPageResponseIndexInfoDto;
 import org.codeiteam3.findex.indexinfo.dto.data.IndexInfoDto;
 import org.codeiteam3.findex.indexinfo.dto.data.IndexInfoSummaryDto;
